refactor: split main.ts into focused modules
Moves the parsers into src/parse/, container lifecycle into
src/container/{create,revision,access}.ts, token exchange into
src/auth.ts, Lockbox version resolution into src/lockbox.ts, and input
reading into readInputs() beside the ActionInputs type it produces.
main.ts keeps credential resolution and orchestration.

Pure movement - the characterization snapshot is unchanged. The public
truthiness quirk (public: 'false' makes the container public) is
preserved deliberately; changing it is a behavior change, not a refactor.
