+++ __tests__/lockbox-secrets.test.ts
@@ -7,6 +7,8 @@ import * as axios from '../__fixtures__/axios.js'
 import * as sdk from '../__fixtures__/yandex-sdk/index.js'
 import { containerService } from '../__fixtures__/yandex-sdk/serverless-containers-v1.js'
 import { __setGetSecretFail, __setSecretList, secretService } from '../__fixtures__/yandex-sdk/lockbox-v1.js'
+// Type-only: erased at compile time, so it does not load src/parse/ before the mocks below.
+import type { Secret as AppSecret } from '../src/parse/index.js'
 
 jest.unstable_mockModule('@actions/core', () => core)
 jest.unstable_mockModule('@actions/github', () => github)
@@ -15,10 +17,11 @@ jest.unstable_mockModule('@yandex-cloud/nodejs-sdk', () => sdk)
 jest.unstable_mockModule('@yandex-cloud/nodejs-sdk/serverless-containers-v1', () => ({ containerService }))
 jest.unstable_mockModule('@yandex-cloud/nodejs-sdk/lockbox-v1', () => ({ secretService }))
 
-const { resolveLatestLockboxVersions, parseLockboxSecretDefinition, parseLockboxVariablesMapping } =
-    await import('../src/main.js')
+// Imported dynamically, after the mock registrations above. A static import would be linked
+// before them, binding src/parse/lockbox-variables.ts to the real @actions/core.
+const { parseLockboxSecretDefinition, parseLockboxVariablesMapping } = await import('../src/parse/index.js')
+const { resolveLatestLockboxVersions } = await import('../src/lockbox.js')
 const { Session } = sdk
-type AppSecret = import('../src/main.js').Secret
 
 describe('resolveLatestLockboxVersions', () => {
     beforeEach(() => {
