+++ __tests__/log-options-min-level.test.ts
@@ -2,7 +2,7 @@ import { expect, test } from '@jest/globals'
 import { getInput } from '@actions/core'
 import { LogLevel_Level } from '@yandex-cloud/nodejs-sdk/dist/generated/yandex/cloud/logging/v1/log_entry'
 
-import { parseLogOptionsMinLevel } from '../src/log-options-min-level.js'
+import { parseLogOptionsMinLevel } from '../src/parse/log-options-min-level.js'
 
 const levelsArray = ['LEVEL_UNSPECIFIED', 'TRACE', 'DEBUG', 'INFO', 'WARN', 'ERROR', 'FATAL'] as Array<
     keyof typeof LogLevel_Level
